build: At least honor the LDFLAGS from the environment
When not honoring the environment we end up linking it without the
GNU_HASH for symbol look-up. Put the content into LIBS and place
these options last to make it link.

We still override the CFLAGS but that seems to not cause issues yet

Fixes:

ERROR: ... No GNU_HASH in the elf binary: 'rtl8168-eeprom' [ldflags]